third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-template-element/definitions/template-contents-owner-document-type-expected.txt

This is a testharness.js-based test.
[FAIL] The template contents owner document type is HTML document (case when document has browsing context and the template is created by HTML parser)
  assert_class_string: Template content owner should be a document expected "[object Document]" but got "[object HTMLDocument]"
[FAIL] The template contents owner document type is HTML document (case when document has browsing context and the template is created by createElement())
  assert_class_string: Template content owner should be a document expected "[object Document]" but got "[object HTMLDocument]"
[FAIL] The template contents owner document type is HTML document (case when document has no browsing context and the template is created by createElement())
  assert_class_string: Template content owner should be a document expected "[object Document]" but got "[object HTMLDocument]"
[FAIL] The template contents owner document type is HTML document (case when document has no browsing context and the template is created via innerHTML)
  assert_class_string: Template content owner should be a document expected "[object Document]" but got "[object HTMLDocument]"
Harness: the test ran to completion.
